// packages/oauth2-server/lib/exceptions/types/supported-oauth2-errors.ts
/**
 * OAuth 2.0 Error Codes supported by Guarani.
 */
export type SupportedOAuth2Error =
  | 'access_denied'
  | 'invalid_client'
  | 'invalid_grant'
  | 'invalid_request'
  | 'invalid_scope'
  | 'server_error'
  | 'temporarily_unavailable'
  | 'unauthorized_client'
  | 'unsupported_grant_type'
  | 'unsupported_response_type';